refactor(tables): rename M3uTableUtils to tableSortingUtils and update imports
- Replaced references to M3uTableUtils with tableSortingUtils in EPGsTable, M3UsTable, UsersTable, and their respective test files.
- Deleted the M3uTableUtils file as its functionality has been consolidated into tableSortingUtils.
- Updated import paths in relevant components and tests to reflect the new utility file name.
