+++ frontend/src/components/tables/tableSortingUtils.jsx
@@ -14,9 +14,8 @@ export const makeHeaderCellRenderer = (sorting, onSortingChange) => (header) => 
   }
 
   return (
-    // nowrap keeps the sort control on the header's line: with the default
-    // wrap, a column only a pixel or two too narrow drops the icon onto a
-    // second row and doubles the header height.
+    // nowrap: default wrap drops the sort icon onto a second row when the
+    // column is only a pixel or two too narrow, doubling the header height.
     <Group gap="xs" wrap="nowrap">
       <Text size="sm" name={header.id}>
         {header.column.columnDef.header}
@@ -43,16 +42,15 @@ export const makeSortingChangeHandler =
       if (sortDirection === false) {
         newSorting[0] = { id: column, desc: true };
       }
-      // third click → clear (empty array)
+      // third click clears sort
     } else {
       newSorting[0] = { id: column, desc: false };
     }
 
     setSorting(newSorting);
-    // Optional: tables that keep their rows in state re-sort them here. Tables
-    // that derive their rows from `sorting` don't pass a callback.
+    // Optional: tables that keep rows in state re-sort here. Tables that
+    // derive rows from `sorting` omit the callback.
     if (onDataSort && newSorting.length > 0) {
       onDataSort(newSorting[0].id, newSorting[0].desc);
     }
   };
-
